+++ PKHeX.Core/PKM/Shared/IShadowPKM.cs
@@ -6,7 +6,7 @@ namespace PKHeX.Core
     /// <remarks>Used only for Colosseum/XD <see cref="PKM"/> that were shadow encounters.</remarks>
     public interface IShadowPKM
     {
-        int ShadowID { get; set; }
+        ushort ShadowID { get; set; }
         int Purification { get; set; }
 
         bool IsShadow { get; }
